requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2==2.9.1
python3-openid==3.2.0
